Retrieve backend info
System info is now being collected and sent to the front end correctly. In order to retrieve the gpu info a new .js file was needed that has more permissions than the system store. In order to read the cpu temp hwinfo needs to be running and logging the temps to a csv file in the users documents which is then read and sent to the front end